==> R_scripts/01_data_aggregation.R <==
## This script is associated with an analysis of spatially and 
## and temporally heterogeneous nutrient loading into Flathead
## Lake's nearshore zone. This script takes data from a cleaned,
## disaggregated format prepares them for an analysis-ready format. Questions about 
##this script should be directed to 
## Michael F. Meyer (michael.f.meyer@wsu.edu). 
## The script can be divided into the following sections:
## 1. Load Packages
## 2. Ash Free Dry Mass
## 3. Nutrients
## 4. Pharmaceuticals and Personal Care Products (PPCPs)
## 5. Stoichiometry
## 6. State Park Attendance
## 7. Fatty Acids
## 8. Temporally Scaled Inverse Distance Weighted Population

# 1. Load Packages --------------------------------------------------------

library(tidyverse)
library(lubridate)
library(stringi)
library(janitor)

# 2. Ash Free Dry Mass ----------------------------------------------------

afdm_orig <- read.csv("../cleaned_disaggregated_data/afdm.csv", header = TRUE)

afdm_clean <- afdm_orig 

head(afdm_orig)

# The data appear to be properly structured for analysis. 

write.csv(x = afdm_clean, 
          file = "../cleaned_data/afdm.csv", 
          row.names = FALSE)


# 3. Nutrients ------------------------------------------------------------

nutrients_orig <- read.csv("../cleaned_disaggregated_data/nutrients.csv", header = TRUE)

nutrients_cleaned <- nutrients_orig %>%
  select(-collection_date) 
  
write.csv(x = nutrients_cleaned, 
          file = "../cleaned_data/nutrients.csv", 
          row.names = FALSE)

# 4. Pharmaceuticals and Personal Care Products (PPCPs) -------------------

ppcp_orig <- read.csv(file = "../cleaned_disaggregated_data/ppcp.csv", header= TRUE)

head(ppcp_orig)

ppcp_cleaned <- ppcp_orig

write.csv(ppcp_cleaned, "../cleaned_data/ppcp.csv", row.names = FALSE)

# 5. Stoichiometry --------------------------------------------------------

stoich_orig <- read.csv(file = "../cleaned_disaggregated_data/stoichiometry.csv",
                        header = TRUE)

stoich_cleaned <- stoich_orig %>%
  select(site, month, stt, tourist_season,
         total_dry_mass_carbon_mg:phosphorus_ug, collection_data_formatted)

write.csv(x = stoich_cleaned, 
          file = "../cleaned_data/stoichiometry.csv", 
          row.names = FALSE)


# 6. Fatty Acids ----------------------------------------------------------

fatty_acids_orig <- read.csv(file = "../cleaned_disaggregated_data/fatty_acids.csv", 
                             header = TRUE)

head(fatty_acids_orig)

write.csv(x = fatty_acids_orig, 
          file = "../cleaned_data/fatty_acids.csv", 
          row.names = FALSE)

# 7. Temporally Scaled Inverse Distance Weighted Population ---------------

# First create a variant where scalars are left in their interpolated states

tsidw_pop <- read.csv(file = "../cleaned_disaggregated_data/temporally_scaled_inverse_distance_weighted_population_metrics.csv",
                      header = TRUE)

head(tsidw_pop)

write.csv(x = tsidw_pop,
          file = "../cleaned_data/temporally_scaled_inverse_distance_weighted_population_metrics.csv",
          row.names = FALSE)

# Second create a varaint with scalars averaged for a given month

tsidw_pop_average <- read.csv(file = "../cleaned_disaggregated_data/averaged_temporally_scaled_inverse_distance_weighted_population_metrics.csv", 
                              header = TRUE)

head(tsidw_pop_average)

tsidw_pop_average_clean <- tsidw_pop_average

write.csv(x = tsidw_pop_average_clean, 
          file = "../cleaned_data/averaged_temporally_scaled_inverse_distance_weighted_population_metrics.csv", 
          row.names = FALSE)

# 8. Periphyton Abundance -------------------------------------------------

periphyton <- read.csv(file = "../cleaned_disaggregated_data/periphyton_abundance.csv",
                       header = TRUE)

head(periphyton)

periphyton_clean <- periphyton

write.csv(x = periphyton_clean, 
          file = "../cleaned_data/periphyton_abundance.csv",
          row.names = FALSE)
